feat(sockets): 채팅 네임스페이스 리팩토링 및 스키마 추가
