+++ src/v1/sockets/chat/chat.schema.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod';
+
+export const joinRoomSchema = z.object({
+  roomId: z.string(),
+});
+
+export const messageSchema = z.object({
+  roomId: z.string(),
+  message: z.string(),
+});
